[NP-50] remove confirm booking modal
